[bugfix]: checkout lvim to branch master
there are some errors this morning, saying there are invalid nodes in treesitter.
Then I run `:checkhealth`, the treesitter gives out a lot of errors.
I guess this is because the core plugin "treesitter` haven't been update for a long time however, I update rust parser yesterday and this must caused some confilicts.
The only resolution now is to use master branch, there are a lot of change in this future release-1.4, so some modification must be done.